Moved files to tidy file system. Sequelize and basic Dog model is working.
