+++ db/schema/walks.sql
