+++ server.js
@@ -7,13 +7,14 @@ const express = require("express");
 const morgan = require("morgan");
 const db = require("./db");
 const app = express();
+// const models = require('../db/models')
 
 
 // const fs = require("fs");
 // const path = require("path");
 const bodyParser = require("body-parser");
 // const helmet = require("helmet");
- const cors = require("cors");
+const cors = require("cors");
 
 
 //Test Sequelize connection
@@ -26,8 +27,8 @@ db.authenticate()
   })
 
   //Sync Models
-  db.sync({ force: true });
-  console.log("All models were synchronized successfully.");
+  // db.sync({force: true})
+  //   console.log("All models were synchronized successfully.");
 
 // Load the logger first so all (static) HTTP requests are logged to STDOUT
 // 'dev' = Concise output colored by response status for development use.
@@ -36,7 +37,7 @@ app.use(morgan("dev"));
 app.use(express.urlencoded({ extended: true }));
 app.use(cors());
 app.use(bodyParser.json())
-  //app.use(express.static("public"));
+//app.use(express.static("public"));
 
 
 
@@ -61,9 +62,10 @@ app.get('*', (req, res) => {
   res.redirect(404, "/");
 });
 
+
+
+
+
 app.listen(PORT, () => {
   console.log(`dogsWalk-api app listening on port ${PORT}`);
 });
-
-
-
